+++ test/functional-api/communications/community-updates/updates.it-spec.ts
@@ -4,20 +4,18 @@ import {
 } from '@test/functional-api/journey/space/space.request.params';
 import { deleteOrganizationCodegen } from '@test/functional-api/organization/organization.request.params';
 import { TestUser } from '@test/utils/token.helper';
-import {
-  delay,
-  entitiesId,
-} from '@test/functional-api/zcommunications/communications-helper';
 import { uniqueId } from '@test/utils/mutations/create-mutation';
 import { changePreferenceSpaceCodegen } from '@test/utils/mutations/preferences-mutation';
 import { users } from '@test/utils/queries/users-data';
-import { assignCommunityRoleToUserCodegen } from '@test/functional-api/integration/community/community.request.params';
 import { createOrgAndSpaceCodegen } from '@test/utils/data-setup/entities';
 import { CommunityRole, SpacePreferenceType } from '@alkemio/client-lib';
 import {
   removeMessageOnRoomCodegen,
   sendMessageToRoomCodegen,
 } from '../communication.params';
+import { entitiesId } from '@test/functional-api/roles/community/communications-helper';
+import { assignCommunityRoleToUserCodegen } from '@test/functional-api/roles/roles-request.params';
+import { delay } from '@test/utils';
 const organizationName = 'upd-org-name' + uniqueId;
 const hostNameId = 'upd-org-nameid' + uniqueId;
 const spaceName = 'upd-eco-name' + uniqueId;
