+++ test/functional-api/conversions/conversion.it-spec.ts
@@ -1,4 +1,3 @@
-import { entitiesId, users } from '../zcommunications/communications-helper';
 import {
   getSpacesCount,
   deleteSpaceCodegen,
@@ -23,11 +22,12 @@ import {
   getChallengeDataCodegen,
 } from '../journey/challenge/challenge.request.params';
 import { uniqueId } from '@test/utils/mutations/create-mutation';
+import { CommunityRole } from '@alkemio/client-lib';
+import { entitiesId, users } from '../roles/community/communications-helper';
 import {
   assignCommunityRoleToOrganizationCodegen,
   assignCommunityRoleToUserCodegen,
-} from '../integration/community/community.request.params';
-import { CommunityRole } from '@alkemio/client-lib';
+} from '../roles/roles-request.params';
 
 const organizationName = 'conv-org-name' + uniqueId;
 const hostNameId = 'conv-org-nameid' + uniqueId;
