+++ test/functional-api/notifications/messaging/mention-user.it-spec.ts
@@ -1,7 +1,6 @@
 /* eslint-disable prettier/prettier */
 import { deleteMailSlurperMails } from '@test/utils/mailslurper.rest.requests';
 import { delay } from '@test/utils/delay';
-import { entitiesId, getMailsData } from '@test/functional-api/zcommunications/communications-helper';
 import { TestUser } from '@test/utils';
 import { uniqueId } from '@test/utils/mutations/create-mutation';
 import { deleteSpaceCodegen } from '@test/functional-api/journey/space/space.request.params';
@@ -21,6 +20,7 @@ import { deleteOrganizationCodegen } from '@test/functional-api/organization/org
 import { UserPreferenceType } from '@alkemio/client-lib';
 import { changePreferenceUserCodegen } from '@test/utils/mutations/preferences-mutation';
 import { sendMessageToRoomCodegen } from '@test/functional-api/communications/communication.params';
+import { entitiesId, getMailsData } from '@test/functional-api/roles/community/communications-helper';
 
 const organizationName = 'urole-org-name' + uniqueId;
 const hostNameId = 'urole-org-nameid' + uniqueId;
