+++ test/functional-api/pagination/organization.it-spec.ts
@@ -167,7 +167,8 @@ describe('Pagination - organization', () => {
     );
   });
 
-  describe('Pagination with cursors', () => {
+  // Skip test due to bug: #3571
+  describe.skip('Pagination with cursors', () => {
     // Arrange
     let startCursor = '';
     let endCursor = '';
