+++ test/functional-api/upload/private-space-document-auth.it-spec.ts
@@ -12,7 +12,6 @@ import {
 import path from 'path';
 import { deleteOrganizationCodegen } from '../organization/organization.request.params';
 import { createOrgAndSpaceWithUsersCodegen } from '@test/utils/data-setup/entities';
-import { entitiesId } from '../zcommunications/communications-helper';
 import { lookupProfileVisuals } from '../lookup/lookup-request.params';
 import { deleteSpaceCodegen } from '../journey/space/space.request.params';
 import {
@@ -23,7 +22,6 @@ import {
   sorted__create_read_update_delete_grant_fileUp_fileDel_contribute,
   sorted__create_read_update_delete_grant_fileUp_fileDel_contribute_updateContent,
 } from '@test/non-functional/auth/my-privileges/common';
-import { createReferenceOnProfileCodegen } from '../integration/references/references.request.params';
 import {
   createLinkCollectionCalloutCodegen,
   createLinkOnCalloutCodegen,
@@ -48,6 +46,8 @@ import { createWhiteboardCalloutCodegen } from '../callout/whiteboard/whiteboard
 import { createWhiteboardRtCalloutCodegen } from '../callout/whiteboardRt/whiteboardRt-callout.params.request';
 import { updateSpacePlatformSettingsCodegen } from '../platform/platform.request.params';
 import { SpaceVisibility } from '@alkemio/client-lib/dist/types/alkemio-schema';
+import { createReferenceOnProfileCodegen } from '../references/references.request.params';
+import { entitiesId } from '../roles/community/communications-helper';
 
 const organizationName = 'org-name' + uniqueId;
 const hostNameId = 'org-nameid' + uniqueId;
