+++ test/functional-api/user-management/invitations/invitation.request.params.ts
@@ -1,10 +1,5 @@
 import { getGraphqlClient } from '@test/utils/graphqlClient';
-import {
-  invitationData,
-  invitationDataExternal,
-  lifecycleData,
-} from '../../../utils/common-params';
-import { graphqlRequestAuth } from '../../../utils/graphql.request';
+import { lifecycleData } from '../../../utils/common-params';
 import { TestUser } from '../../../utils/token.helper';
 import { graphqlErrorWrapper } from '@test/utils/graphql.wrapper';
 
@@ -67,22 +62,6 @@ export const inviteExternalUserCodegen = async (
   return graphqlErrorWrapper(callback, userRole);
 };
 
-// export const removeInvitation = async (appId: string) => {
-//   const requestParams = {
-//     operationName: null,
-//     query: `mutation deleteInvitation($deleteData: DeleteInvitationInput!) {
-//       deleteInvitation(deleteData: $deleteData) {
-//         ${invitationData}}}`,
-//     variables: {
-//       deleteData: {
-//         ID: appId,
-//       },
-//     },
-//   };
-
-//   return await graphqlRequestAuth(requestParams, TestUser.GLOBAL_ADMIN);
-// };
-
 export const deleteInvitationCodegen = async (
   invitationId: string,
   userRole: TestUser = TestUser.GLOBAL_ADMIN
@@ -121,22 +100,6 @@ export const deleteExternalInvitationCodegen = async (
   return graphqlErrorWrapper(callback, userRole);
 };
 
-// export const removeExternalInvitation = async (appId: string) => {
-//   const requestParams = {
-//     operationName: null,
-//     query: `mutation deleteExternalInvitation($deleteData: DeleteInvitationExternalInput!) {
-//       deleteInvitationExternal(deleteData: $deleteData) {
-//         ${invitationDataExternal}}}`,
-//     variables: {
-//       deleteData: {
-//         ID: appId,
-//       },
-//     },
-//   };
-
-//   return await graphqlRequestAuth(requestParams, TestUser.GLOBAL_ADMIN);
-// };
-
 export const getSpaceInvitationCodegen = async (
   spaceId: string,
   userRole: TestUser = TestUser.GLOBAL_ADMIN
@@ -153,17 +116,3 @@ export const getSpaceInvitationCodegen = async (
     );
   return graphqlErrorWrapper(callback, userRole);
 };
-
-// export const getExternalInvitation = async (
-//   spaceNameId: string,
-//   userRole: TestUser = TestUser.NON_HUB_MEMBER
-// ) => {
-//   const requestParams = {
-//     operationName: null,
-//     variables: {},
-//     query: `query{space(ID: "${spaceNameId}" ) {community{
-//       invitationsExternal{${invitationDataExternal}}}}}`,
-//   };
-
-//   return await graphqlRequestAuth(requestParams, userRole);
-// };
