+++ test/non-functional/auth/my-privileges/entity-based/private-hub/private-challenge.it-spec.ts
@@ -5,7 +5,6 @@ import {
   getSpaceDataCodegen,
   deleteSpaceCodegen,
 } from '@test/functional-api/journey/space/space.request.params';
-import { entitiesId } from '@test/functional-api/zcommunications/communications-helper';
 import {
   readPrivilege,
   sorted__create_read_update_delete_grant_createDiscussion_Privilege,
@@ -38,6 +37,7 @@ import {
   createOrgAndSpaceWithUsersCodegen,
 } from '@test/utils/data-setup/entities';
 import { SpacePreferenceType } from '@alkemio/client-lib';
+import { entitiesId } from '@test/functional-api/roles/community/communications-helper';
 
 const organizationName = 'ch-pref-org-name' + uniqueId;
 const hostNameId = 'ch-pref-org-nameid' + uniqueId;
